+++ server/core/src/tide/manifest.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 ///! Builds a Progressive Web App Manifest page.
 // Thanks to the webmanifest crate for a lot of this code
-use crate::https::{AppState, RequestExtensions};
+use crate::tide::{AppState, RequestExtensions};
 
 /// The MIME type for `.webmanifest` files.
 const MIME_TYPE_MANIFEST: &str = "application/manifest+json;charset=utf-8";
